Single-sourcing local and e2e tests for validation
