+++ .claspignore
@@ -3,7 +3,6 @@
 dist/**
 docs/**
 node_modules/**
-tests/**
 src/*.test.ts
 src/index.ts
 !appsscript.json
